# techforce-pro/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: FieldCommand — Premium fire protection field service management platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List jobs with optional filters
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: employeeId
          in: query
          schema:
            type: ["integer", "null"]
        - name: customerId
          in: query
          schema:
            type: ["integer", "null"]
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"
    post:
      operationId: createJob
      tags: [jobs]
      summary: Create job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /jobs/{id}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get job by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
    patch:
      operationId: updateJob
      tags: [jobs]
      summary: Update job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
    delete:
      operationId: deleteJob
      tags: [jobs]
      summary: Delete job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Job:
      type: object
      required: [id, customerId, customerName, employeeId, techName, serviceType, scheduledDate, scheduledTime, status, priority, revenue, certRequired]
      properties:
        id:
          type: integer
        customerId:
          type: integer
        customerName:
          type: string
        employeeId:
          type: ["integer", "null"]
        techName:
          type: string
        serviceType:
          type: string
        description:
          type: string
        scheduledDate:
          type: string
        scheduledTime:
          type: string
        status:
          type: string
          enum: [pending, in_progress, completed, return, reschedule, emergency, cancelled, non_compliant, will_return]
        priority:
          type: string
          enum: [low, medium, high, critical]
        revenue:
          type: number
        certRequired:
          type: string
        recurring:
          type: boolean
        recurringInterval:
          type: ["string", "null"]
        notes:
          type: string
        completedAt:
          type: ["string", "null"]
        invoiceId:
          type: ["integer", "null"]
        estimateId:
          type: ["integer", "null"]
        nonComplianceReason:
          type: ["string", "null"]
        nonComplianceNotifiedAt:
          type: ["string", "null"]
    JobInput:
      type: object
      required: [customerId, serviceType, scheduledDate, scheduledTime, revenue]
      properties:
        customerId:
          type: integer
        employeeId:
          type: ["integer", "null"]
        serviceType:
          type: string
        description:
          type: string
        scheduledDate:
          type: string
        scheduledTime:
          type: string
        status:
          type: string
        priority:
          type: string
        revenue:
          type: number
        certRequired:
          type: string
          nullable: true
        recurring:
          type: boolean
        recurringInterval:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        estimateId:
          type: ["integer", "null"]
    JobUpdate:
      type: object
      properties:
        employeeId:
          type: ["integer", "null"]
        serviceType:
          type: string
        description:
          type: string
        scheduledDate:
          type: string
        scheduledTime:
          type: string
        status:
          type: string
        priority:
          type: string
        revenue:
          type: number
        notes:
          type: string
        nonComplianceReason:
          type: ["string", "null"]
...